sql backed register; calendar refactor
So I didn't like the Init() approach on Calendar. Keeps main simpler
and the book/nil checks to create the table and prepared statements
on demand works well. It's nice to have the query close to the scan.
I'd rather not have to call createTableOnce all the time, but maybe
there's something better out there.