form.Submit() in a transaction
Form.Submit() is three operations: register/charge/book.
For now, the best I can come up with is to pass around
a transaction that Submit() manages. This creates dual
functions like List/ListTx to support both complete and
partial usage. It may not be necessary to support the
non-Tx call if you detect nil Tx and boot one. But, that
puts undue usage burden on the caller. The two functions
while somewhat duplicative are unambiguous in their use.
Perhaps the Begin/Commit/Rollback boilerplate can be
lessened.